Deduplicate test helpers into shared TestHelpers.swift
Extract identical makeEntityJSON and stubEntityURL helpers from both
test files into module-level free functions, and add a stubEntityURL(data:)
overload so all URL-building logic lives in one place.
